arm64/rusage: fix ihk_pgsizes type
1ULL << 42 is too big to fit into an int

+++ linux/include/ihk/arch/arm64/ihk/ihk_arch_rusage.h
@@ -17,7 +17,7 @@
 # define GRANULE_SIZE _SZ4KB
 #endif
 
-const int ihk_pgsizes[] = {
+const long ihk_pgsizes[] = {
 #if GRANULE_SIZE == _SZ4KB
 	1ULL << 12,
 	1ULL << 21,
